feat: ✨ 支持大位移并提示异速往返误配
- 单个 HID 报告的 X/Y 仅支持 -127–127，采用 schema v1 兼容的有符号 10 位扩展与自动拆包满足 -500–500 总位移
- 不扩大动作记录，避免旧 Flash 配置失效并保持现有配置长度和分块协议
- 异速往返会触发主机指针加速，网页主动警告而不采用不可移植的屏幕像素补偿
- 详见 docs/protocol.md §配置 schema v1

+++ include/hidpilot_executor.h
@@ -43,8 +43,8 @@ typedef struct {
     uint16_t move_step_count;
     uint16_t move_step_index;
     uint8_t action_index;
-    int8_t move_target_x;
-    int8_t move_target_y;
+    int16_t move_target_x;
+    int16_t move_target_y;
     int8_t move_target_wheel;
     int8_t move_target_pan;
     bool mounted;
